Working UI Finished (#1)
* Uploaded data

Created the first draft of the data graphing

* Working UI Finished

The user can now access all types of different graphs by warehouse or by category of purchase. On the bottom of the website, there is a statement of interest for the users to see what the candidate's stance is on the data.